requirements.txt
noobstuffs
PyGithub
python-telegram-bot
